+++ hydxlib/__init__.py
@@ -1,0 +1,3 @@
+# fmt: off
+__version__ = '0.8.dev0'
+# fmt: on