+++ hydxlib/tests/test_exporter.py
@@ -1,34 +1,13 @@
 # -*- coding: utf-8 -*-
 """Tests for importer.py"""
-from unittest import TestCase
-import pytest
-
-from hydxlib.sql_models.threedi_database import ThreediDatabase
+from hydxlib.exporter import export_threedi
+from hydxlib.exporter import get_cross_section_definition_id
+from hydxlib.exporter import get_start_and_end_connection_node
+from hydxlib.exporter import write_threedi_to_db
 from hydxlib.importer import import_hydx
 from hydxlib.threedi import Threedi
-from hydxlib.exporter import (
-    export_threedi,
-    write_threedi_to_db,
-    get_cross_section_definition_id,
-    get_start_and_end_connection_node,
-)
 
-
-def test_check_connection_db(caplog):
-    # temporarily db setup!
-    db = ThreediDatabase(
-        {
-            "host": "localhost",
-            "port": "5432",
-            "database": "test_gwsw",
-            "username": "postgres",
-            "password": "postgres",
-        },
-        "postgres",
-    )
-
-    session = db.get_session()
-    assert session is not None
+import pytest
 
 
 def test_get_start_and_end_connection_node_right():
@@ -52,39 +31,31 @@ def test_get_cross_section_definition_id_wrong(caplog):
     assert "Cross section" in caplog.text
 
 
-class TestThreedi(TestCase):
-    def setUp(self):
-        self.threedi = Threedi()
-        hydx_path = "hydxlib/tests/example_files_structures_hydx/"
-        self.threedi_db_settings = {
-            "threedi_dbname": "test_gwsw",
-            "threedi_host": "localhost",
-            "threedi_user": "postgres",
-            "threedi_password": "postgres",
-            "threedi_port": 5432,
-        }
-        self.hydx = import_hydx(hydx_path)
-        self.threedi.import_hydx(self.hydx)
+@pytest.fixture(scope="session")
+def hydx_setup():
+    threedi = Threedi()
+    hydx_path = "hydxlib/tests/example_files_structures_hydx/"
+    hydx = import_hydx(hydx_path)
+    threedi.import_hydx(hydx)
+    return hydx, threedi
 
-    @pytest.fixture(autouse=True)
-    def inject_fixtures(self, caplog):
-        self._caplog = caplog
 
-    def test_export_threedi(self):
-        output = export_threedi(self.hydx, self.threedi_db_settings)
-        assert len(output.connection_nodes) == 85
+def test_export_threedi(hydx_setup, mock_exporter_db):
+    output = export_threedi(hydx_setup[0], "/some/path")
+    assert len(output.connection_nodes) == 85
 
-    def test_write_to_db_con_nodes_huge(self):
-        commit_counts_expected = {
-            "connection_nodes": 85,
-            "manholes": 84,
-            "pumpstations": 8,
-            "weirs": 6,
-            "cross_sections": 38,
-            "orifices": 2,
-            "impervious_surfaces": 330,
-            "pipes": 80,
-            "outlets": 3,
-        }
-        commit_counts = write_threedi_to_db(self.threedi, self.threedi_db_settings)
-        assert commit_counts == commit_counts_expected
+
+def test_write_to_db_con_nodes_huge(hydx_setup, mock_exporter_db):
+    commit_counts_expected = {
+        "connection_nodes": 85,
+        "manholes": 84,
+        "pumpstations": 8,
+        "weirs": 6,
+        "cross_sections": 41,
+        "orifices": 2,
+        "impervious_surfaces": 330,
+        "pipes": 80,
+        "outlets": 3,
+    }
+    commit_counts = write_threedi_to_db(hydx_setup[1], "/some/path")
+    assert commit_counts == commit_counts_expected